check for existing active FQDN

+++ providers/computername.rb
@@ -29,24 +29,35 @@ end
 
 action :add_primary do
 
-  converge_by("add new primary Fully Qualified Domain Name #{new_resource.host} for computer #{new_resource.name}") do
+  unless new_resource.host.casecmp( activeFqdn() )
 
-    reboot "New ComputerName" do
-      action :nothing
-      reason "The new primary computername #{new_resource.host} needs a reboot to become effective."
+    converge_by("add new primary Fully Qualified Domain Name #{new_resource.host} for computer #{new_resource.name}") do
+  
+      reboot "New ComputerName" do
+        action :nothing
+        reason "The new primary computername #{new_resource.host} needs a reboot to become effective."
+      end
+      
+      powershell_script "netdom computername" do
+        code <<-EOH
+          netdom computername #{new_resource.name} /add:#{new_resource.host}
+          netdom computername #{new_resource.name} /makeprimary:#{new_resource.host}
+        EOH
+        
+        notifies new_resource.reboot_immediately ? :reboot_now : :request_reboot, 'reboot[New ComputerName]', :immediately
+      end
+  
+      Chef::Log.info("#{new_resource} added primary #{new_resource.host} for #{new_resource.name}")
     end
     
-    powershell_script "netdom computername" do
-      code <<-EOH
-        netdom computername #{new_resource.name} /add:#{new_resource.host}
-        netdom computername #{new_resource.name} /makeprimary:#{new_resource.host}
-      EOH
-      
-      notifies new_resource.reboot_immediately ? :reboot_now : :request_reboot, 'reboot[New ComputerName]', :immediately
-    end
-
-    Chef::Log.info("#{new_resource} added primary #{new_resource.host} for #{new_resource.name}")
+    new_resource.updated_by_last_action(true)
   end
+end
 
-  new_resource.updated_by_last_action(true)
+def activeFqdn()
+  
+  hostname = registry_get_values( "HKLM\SYSTEM\CurrentControlSet\Control\ComputerName\ActiveComputerName" )['ComputerName']
+  domain   = registry_get_values( "HKLM\SYSTEM\CurrentControlSet\Services\Tcpip\Parameters" )['Domain']
+  
+  "#{hostname}.#{domain}"  
 end
